Actualizada la interfaz python-c, conseguida eficiencia optima frente a los metodos set y keyvalues de dictionary. El metodo implementado vuelve a encargarse del memset del aux, no forma parte de los datos necesarios para transpasar los datos de python a c o viceversa y no parecia justo descontarse ese tiempo.

+++ src/C/wrapper.h
@@ -1,5 +1,5 @@
 
-int wrapper(float *vin,    // Vector de entrada
-            float *vout,         // Vector de salida
-            int size);     // Longitud de los vectores
+int wrapper(int *vin,    // Vector de entrada
+            int *vout,         // Vector de salida
+            int size);       // Longitud de los vectores
